frontend/api/requirements.txt
openai==1.54.0
chromadb==0.4.22
sentence-transformers==2.3.1
python-dotenv==1.0.0
